# Conflicts:
#	api/pom.xml
#	api/src/main/java/pro/beerpong/api/control/GroupController.java
#	api/src/main/java/pro/beerpong/api/control/MatchController.java
#	api/src/main/java/pro/beerpong/api/control/ProfileController.java
#	api/src/main/java/pro/beerpong/api/mapping/PlayerMapper.java
#	api/src/main/java/pro/beerpong/api/mapping/ProfileMapper.java
#	api/src/main/java/pro/beerpong/api/model/dao/Rule.java
#	api/src/main/java/pro/beerpong/api/model/dao/RuleMove.java
#	api/src/main/java/pro/beerpong/api/model/dto/ErrorCodes.java
#	api/src/main/java/pro/beerpong/api/model/dto/GroupCreateDto.java
#	api/src/main/java/pro/beerpong/api/model/dto/PlayerDto.java
#	api/src/main/java/pro/beerpong/api/model/dto/ResponseEnvelope.java
#	api/src/main/java/pro/beerpong/api/repository/GroupRepository.java
#	api/src/main/java/pro/beerpong/api/repository/ProfileRepository.java
#	api/src/main/java/pro/beerpong/api/service/GroupService.java
#	api/src/main/java/pro/beerpong/api/service/MatchService.java
#	api/src/main/java/pro/beerpong/api/service/PlayerService.java
#	api/src/main/java/pro/beerpong/api/service/ProfileService.java
#	api/src/main/java/pro/beerpong/api/service/SeasonService.java
#	docker/docker-compose-dev.yml

+++ api/src/main/java/pro/beerpong/api/service/ProfileService.java
@@ -1,9 +1,8 @@
 package pro.beerpong.api.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import pro.beerpong.api.mapping.ProfileMapper;
-import pro.beerpong.api.model.dao.Profile;
 import pro.beerpong.api.model.dto.ProfileCreateDto;
 import pro.beerpong.api.model.dto.ProfileDto;
 import pro.beerpong.api.repository.GroupRepository;
@@ -13,18 +12,12 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
+@RequiredArgsConstructor
 public class ProfileService {
-
     private final ProfileRepository profileRepository;
     private final GroupRepository groupRepository;
     private final ProfileMapper profileMapper;
-
-    @Autowired
-    public ProfileService(ProfileRepository profileRepository, GroupRepository groupRepository, ProfileMapper profileMapper) {
-        this.profileRepository = profileRepository;
-        this.groupRepository = groupRepository;
-        this.profileMapper = profileMapper;
-    }
+    private final PlayerService playerService;
 
     public ProfileDto createProfile(String groupId, ProfileCreateDto profileCreateDto) {
         var groupOptional = groupRepository.findById(groupId);
@@ -33,7 +26,17 @@ public class ProfileService {
         profile.setGroup(groupOptional.orElseThrow());
 
         var savedProfile = profileRepository.save(profile);
+
+        playerService.createPlayer(savedProfile.getGroup().getActiveSeason(), savedProfile);
+
         return profileMapper.profileToProfileDto(savedProfile);
+    }
+
+    public List<ProfileDto> listAllProfilesOfGroup(String groupId) {
+        return profileRepository.findAllByGroupId(groupId)
+                .stream()
+                .map(profileMapper::profileToProfileDto)
+                .collect(Collectors.toList());
     }
 
     public List<ProfileDto> listAllProfiles() {
@@ -49,8 +52,12 @@ public class ProfileService {
                 .orElse(null);
     }
 
-    public void deleteProfile(String id) {
-        profileRepository.deleteById(id);
+    public boolean deleteProfile(String id) {
+        if (profileRepository.existsById(id)) {
+            profileRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
     public ProfileDto updateProfile(String groupId, String id, ProfileCreateDto profileCreateDto) {
